+++ apphelpers/utilities/email.py
@@ -1,19 +1,20 @@
-import html2text
 import os
 import smtplib
 import ssl
-
-from email.utils import formataddr
+from email import encoders
+from email.mime.base import MIMEBase
+from email.mime.image import MIMEImage
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
-from email.mime.image import MIMEImage
-from email.mime.base import MIMEBase
-from email import encoders
+from email.utils import formataddr
 
+import html2text
 from converge import settings
 
+from apphelpers.loggers import app_logger
 
-def send_email(
+
+def format_msg(
     sender,
     recipients,
     subject,
@@ -23,19 +24,8 @@ def send_email(
     images=None,
     reply_to=None,
     bcc=None,
+    headers=None,
 ):
-    """
-    text: text message. If html is provided and not text, text will be auto generated
-    html: html message
-    images: list of cid and image paths.
-        eg. [('logo', 'images/logo.png'), ('Bruce', 'images/bat.png')]
-
-    sender: can be sender email string e.g. 'foo@example.com' or
-    list/tuple sender name and email  ('Foo', 'foo@example.com')
-
-    """
-    assert any((text, html)), "please provide html or text"
-
     if html and not text:
         text = html2text.html2text(html)
 
@@ -72,6 +62,71 @@ def send_email(
                 "Content-Disposition", f"attachment; filename= {file_name}"
             )
             msg.attach(file_part)
+    if headers:
+        for key, value in headers.items():
+            msg.add_header(key, value)
+    return msg
+
+
+def send_email(
+    sender,
+    recipients,
+    subject,
+    text=None,
+    html=None,
+    attachments=None,
+    images=None,
+    reply_to=None,
+    bcc=None,
+    headers=None,
+):
+    """
+    text: text message. If html is provided and not text, text will be auto generated
+    html: html message
+    images: list of cid and image paths.
+        eg. [('logo', 'images/logo.png'), ('Bruce', 'images/bat.png')]
+
+    sender: can be sender email string e.g. 'foo@example.com' or
+    list/tuple sender name and email  ('Foo', 'foo@example.com')
+    headers: Dictionary of additional headers.
+    """
+    assert any((text, html)), "please provide html or text"
+
+    if settings.DEBUG or settings.APP_MODE != "prod":
+        # Make sure that we don't send emails to external emails in dev/stage
+        filtered_recipients = []
+        for recpt in recipients + (bcc or []):
+            if isinstance(recpt, (list, tuple)):
+                _, email = recpt
+            else:
+                email = recpt
+
+            if any(
+                email.endswith(f"@{domain}")
+                for domain in settings.INTERNAL_EMAIL_DOMAINS
+            ):
+                filtered_recipients.append(recpt)
+            else:
+                internal_domains = ", ".join(settings.INTERNAL_EMAIL_DOMAINS)
+                app_logger.info(
+                    f"Skipping email to {email} as it does not end with any of"
+                    f" {internal_domains}"
+                )
+
+        recipients = filtered_recipients
+
+    msg = format_msg(
+        sender,
+        recipients,
+        subject,
+        text,
+        html,
+        attachments,
+        images,
+        reply_to,
+        bcc,
+        headers,
+    )
 
     context = ssl.create_default_context()